+++ Project/WFColorCodeTests/WFColorCodeTests.m
@@ -98,4 +98,19 @@
     XCTAssertNil([NSColor colorWithHex:0xFFFFFF + 1 alpha:1.0]);
 }
 
+
+- (void)testKeyword
+{
+    WFColorCodeType colorCodeType;
+    NSColor *color = [NSColor colorWithColorCode:@"navy" codeType:&colorCodeType];
+    
+    XCTAssertEqual(colorCodeType, WFColorCodeCSSKeyword);
+    XCTAssertEqualObjects([color colorCodeWithType:WFColorCodeCSSKeyword], @"Navy");
+    XCTAssertNil([NSColor colorWithColorCode:@"foobar" codeType:NULL]);
+    
+    NSDictionary *keywordColors = [NSColor stylesheetKeywordColors];
+    NSString *keyword = @"Orange";
+    XCTAssertEqualObjects(keywordColors[keyword], [NSColor colorWithColorCode:keyword codeType:NULL]);
+}
+
 @end